+++ mininn/test/runtime/predictor_test.cc
@@ -47,7 +47,7 @@ TEST(Predictor, get_input_tensors) {
   auto predictor = std::make_shared<Predictor>(graph);
   std::vector<std::shared_ptr<Tensor>> input_tensors =
       predictor->get_input_tensors();
-  int input_num = input_tensors.size();
+  int input_num = static_cast<int>(input_tensors.size());
   EXPECT_EQ(input_num, 2);
 }
 
@@ -57,7 +57,7 @@ TEST(Predictor, get_output_tensors) {
   auto predictor = std::make_shared<Predictor>(graph);
   std::vector<std::shared_ptr<Tensor>> output_tensors =
       predictor->get_output_tensors();
-  int output_num = output_tensors.size();
+  int output_num = static_cast<int>(output_tensors.size());
   EXPECT_EQ(output_num, 1);
 }
